RD-796 Adding deployments labels to include (#2612)
labels are defined in the Deployment table as a backref from the DeploymentLabel table and not as a column. So, putting labels in the _include parameter doesn't work. The solution for this is to get the whole Deployment element and let the _include parameter refine the returned results, according to the values specified in it.