marshal_with: add force_get_data
It is often the case that I want a different representation for
listing, and a different for getting a single object (because
of some expensive-to-compute attributes that shouldn't be present
in the list).

We already have a way to provide `?_get_data=true` to the query
but it'll be useful to just force it, so that the user doesn't
need to always provide it.