Execution scheduler: don't stay in transaction for a long time
When there's no schedules to run, close the transaction.
Otherwise, the process will stay in "idle in transaction" mode for
a long time, keeping table locks and server resources for no reason.

This is only the case when we don't run the schedule, because if
we do, we commit the transaction in sm.update()