RD-1414 Throw NotFound if blueprint is not found in extract (#2688)
This is called in the last "blueprint state update" call of the
upload_blueprint workflow. This is so that, if the blueprint is
downloaded from a url, the restservice knows to unpack the
blueprint archive then.

So if the upload_blueprint workflow hits the node who _doesn't_ have
the blueprint (because it wasn't replicated yet due to replication
lag), it needs to hit the other node as well, to tell that one to
unpack the archive.