RD-1429 Add a timeout to the patroni requests (#2685)
Otherwise, the restservice will stay dead forever, if a DB node
is blackholing traffic.

I used 5 as the timeout because it seems good enough. You'd have
already thrown your holy hand grenade by 3.

Note that (as opposed to the requests in RESTClients), this is
both the connect+read timeout, there's no need to distinguish here.