Pass auto-correct-types to update_plugin workflow
--auto-correct-types flag issued with plugins update should be passed
down to the deploymetns update flows.  This patch is the first step on
this path – it passes this flag to the update_plugin system workflow in
execution_parameters dictionary.  The use of the dictionary and not
another database column is a consequence of the commitment not to modify
the database schema in patch releases (and this patch is intended to be
a part of 5.1.3 release).