RD-1515 Add --auto-correct-types to plugins update (#2710)
* Pass auto-correct-types to update_plugin workflow

--auto-correct-types flag issued with plugins update should be passed
down to the deploymetns update flows.  This patch is the first step on
this path – it passes this flag to the update_plugin system workflow in
execution_parameters dictionary.  The use of the dictionary and not
another database column is a consequence of the commitment not to modify
the database schema in patch releases (and this patch is intended to be
a part of 5.1.3 release).

* Test auto-correct-types flag in plugins update

* Fix rest endpoints test: test_plugins_update_and_execution_parameters_are_correct

* Actually don't test auto-correct-types in integration tests

* Test plugins update auto-correct-types flag in rest endpoint

* In-review improvements

* One more test fix (workflows' test_plugins)