RD-1514 plugins update decoupling (#2720)
* Easier to read temporary blueprint metadata

While updating plugins a temporary blueprint is created.  This patch
makes its `id` and `description` more descriptive, which will help to
find the original blueprint from which the copy was created.

* No RuntimeError on failed deployment update

Do not raise RuntimeError in case one of the deployments to update fails
to be updated.  Instead just send event/log and continue.

* Make manual update or continuation possible

In case there were deployments which failed to be updated during plugin
update flow, make the temporary blueprint visible and log the deployment
IDs for reference.

* Make a blueprint duplicate for updated deployments

If some of the updated deployments were successful and others were not,
let's make a copy of the original blueprint (and its archive).  Updated
deployments will refer to the copy, and unupdated deployments will refer
to the original.

* Fix PluginsUpdateTest for finalize_plugins_update

The solution includes pretending a deployment was updated during plugins
update procedure.

* Add test for partial only plugins update

* Update plugins_update workflow test

Test the workflow for continuation of the update, regardess of
deployment update failure(s).

* Don't use str.format() in logger

* plugins-update-[TIMESTAMP] prefix, not suffix

Adding a prefix to blueprint.id is a cleaner solution.