Plugins and deployment update improvements (#2772)
* CYBL-1107 Inputs type casting functionality (#2702)

In some rare (?) cases there might exist deployments with invalid type
of inputs (e.g. '20' stored where input type should be integer).  This
patch adds functionality to attempt to automatically cast types of input
values to expected (a.k.a. inputs auto-correction).

* RD-1515 Add --auto-correct-types to plugins update (#2710)

* Pass auto-correct-types to update_plugin workflow

--auto-correct-types flag issued with plugins update should be passed
down to the deploymetns update flows.  This patch is the first step on
this path – it passes this flag to the update_plugin system workflow in
execution_parameters dictionary.  The use of the dictionary and not
another database column is a consequence of the commitment not to modify
the database schema in patch releases (and this patch is intended to be
a part of 5.1.3 release).

* Test auto-correct-types flag in plugins update

* Fix rest endpoints test: test_plugins_update_and_execution_parameters_are_correct

* Actually don't test auto-correct-types in integration tests

* Test plugins update auto-correct-types flag in rest endpoint

* In-review improvements

* One more test fix (workflows' test_plugins)

* RD-1514 plugins update decoupling (#2720)

* Easier to read temporary blueprint metadata

While updating plugins a temporary blueprint is created.  This patch
makes its `id` and `description` more descriptive, which will help to
find the original blueprint from which the copy was created.

* No RuntimeError on failed deployment update

Do not raise RuntimeError in case one of the deployments to update fails
to be updated.  Instead just send event/log and continue.

* Make manual update or continuation possible

In case there were deployments which failed to be updated during plugin
update flow, make the temporary blueprint visible and log the deployment
IDs for reference.

* Make a blueprint duplicate for updated deployments

If some of the updated deployments were successful and others were not,
let's make a copy of the original blueprint (and its archive).  Updated
deployments will refer to the copy, and unupdated deployments will refer
to the original.

* Fix PluginsUpdateTest for finalize_plugins_update

The solution includes pretending a deployment was updated during plugins
update procedure.

* Add test for partial only plugins update

* Update plugins_update workflow test

Test the workflow for continuation of the update, regardess of
deployment update failure(s).

* Don't use str.format() in logger

* plugins-update-[TIMESTAMP] prefix, not suffix

Adding a prefix to blueprint.id is a cleaner solution.

* CYBL-1114 Reevaluate plugins update statuses (#2741)

* Extra parameters for status reevaluation of plugins updates

* Reevaluation function implementation

* Fix execution_status type

* CYBL-1114 Reevaluate deployment update statuses (#2753)

* Move no active update running validation

* Pass reevaluate_active_statuses to depl update

* Accept reevaluate_active_statuses flag in depl update

* Reevaluate active depl update statuses

If reevaluate_active_statuses flag is set, update active deployment
statuses based on relevant executions' statuses before validating no
active deployment updates are running for a given deployment_id.

* Fix tests

* RD-1637 Use EXACT_VERSION in update-plugin-imports script (#2755)

* Get rid of unnecessary REPO parameter

* Allow to define EXACT_VERSION

If a plugin has the parameter `EXACT_VERSION` set, it's value will be
always used as a suggested plugin version.