move merge_parameters to the execution model
The method figures out the actual parameters, based on the passed-in
params and the workflow.

Instead of keeping it as a method on resource_manager, instead, put
it as a validator on the execution model itself, so that creating
an execution object already does the merge-and-validate stuff.

That is, doing `Execution(parameters=x, deployment=y, workflow_id=z)`
will by itself run the validations and parameter normalization.
No need for a separate method. This is better because now invalid
state is not representable. And it's impossible to forget to call
the validate method.

Those are `@validates` methods because that's the best way I've found
for a setter-for-relationship. I'd prefer the `deployment` one to
be something else, but it can't be.

The approach, with the multiple `@validates` means: only run the
merge-and-validate method when all 3 (params, dep, wf_id) are set.
Before they're set, don't run it yet.

Similarly for set_deployment.

`allow_custom_parameters` is unfortunate, but it must be passed
in _somewhere_.