make resource_manager.execute_workflow take the execution object
The idea is, I want to split creating the execution object itself,
from sending it. This makes rm.execute_workflow take the execution
object, and, well, executes it.

With this, we note that all the kwargs for `rm.execute_workflow`
are split in two groups: the kwargs for the execution object itself,
which are stored in the db, and the kwargs for sending the execution
(`bypass_maintenance`, `wait_after_fail`, etc), which are sned-time
settings.

Oh, `_set_execution_tenant` is just unnecessary. `workflow_executor`
doesn't use `current_tenant` anymore (or `current_user`,
for that matter)