move setting latest_execution
set .latest_execution only when the execution is actually started.
Same for .started_at.

It used to be somewhere in .execute_workflow before, but I removed
that, because that was worse: it's not really the latest execution,
nor it is in progress, until it is started. It also shouldn't have
a .started_at before it's actually started.