remove stupid test
I have no idea what these tests are about. They create system
workflows, but those aren't really system workflows, because these
aren't valid workflow_ids? And then the test asserts that the execution
exists? What's so special about that execution?

If it's trying to test that you can create system workflow with any
workflow_id, which I'm not sure it does, then no: you can't.
We validate the workflow_id to be on the whitelist, for system
worfklows.