RD-1640 execute_workflow refactor + queueing (#2781)
* move merge_parameters to the execution model

The method figures out the actual parameters, based on the passed-in
params and the workflow.

Instead of keeping it as a method on resource_manager, instead, put
it as a validator on the execution model itself, so that creating
an execution object already does the merge-and-validate stuff.

That is, doing `Execution(parameters=x, deployment=y, workflow_id=z)`
will by itself run the validations and parameter normalization.
No need for a separate method. This is better because now invalid
state is not representable. And it's impossible to forget to call
the validate method.

Those are `@validates` methods because that's the best way I've found
for a setter-for-relationship. I'd prefer the `deployment` one to
be something else, but it can't be.

The approach, with the multiple `@validates` means: only run the
merge-and-validate method when all 3 (params, dep, wf_id) are set.
Before they're set, don't run it yet.

Similarly for set_deployment.

`allow_custom_parameters` is unfortunate, but it must be passed
in _somewhere_.

* implement execution.render_context()

the context that's used in workflow_executor, it would be better
if the execution itself knew how to format it. Move it to here, then.

* rewrite workflow_executor methods

workflow_executor.execute_workflow and .execute_system_workflow
already take the execution object, so this is just the next step -
having the execution object itself render the context to be sent.
This allows us to only have a single execute method, no need for the
execute_workflow/system_workflow split anymore here.

Unfortunately, some things are still populated in the context here,
because they need to be retrieved from the db at send time. I couldn't
get rid of them right now. Maybe someday.

* make resource_manager.execute_workflow take the execution object

The idea is, I want to split creating the execution object itself,
from sending it. This makes rm.execute_workflow take the execution
object, and, well, executes it.

With this, we note that all the kwargs for `rm.execute_workflow`
are split in two groups: the kwargs for the execution object itself,
which are stored in the db, and the kwargs for sending the execution
(`bypass_maintenance`, `wait_after_fail`, etc), which are sned-time
settings.

Oh, `_set_execution_tenant` is just unnecessary. `workflow_executor`
doesn't use `current_tenant` anymore (or `current_user`,
for that matter)

* pass the execution object into rm.execute_workflow

now that I made this method take just the execution object, pass
it in, in all the places that call rm.execute_workflow/system_workflow

* rewrite check-for-executions

just nicer this way.

* rewrite delete-dep-env

also just nicer this way. I note that the python-side checking for
execution status is pointless, since we already filter on the db side.

* rewrite execute_queued_workflow

no need for all that old bullshit. It can really be as simple as
that.
(some later commit does actually just remove this method altogether)

* move setting latest_execution

set .latest_execution only when the execution is actually started.
Same for .started_at.

It used to be somewhere in .execute_workflow before, but I removed
that, because that was worse: it's not really the latest execution,
nor it is in progress, until it is started. It also shouldn't have
a .started_at before it's actually started.

* small renames

just ignore this one.

* add defaults for a few columns

it's way nicer when stuff is not always required.

* add missing imports

also just ignore this one.

* rewrite execution param validation tests

it's way easier when we have the validation just in the object itself:
we can unittest it directly, without going through the whole system,
by just creating the object. No need for DB in these tests, either.

* fix mock use in tests

the method changed, the mock needs to be changed too.

Also, in the exec-group test, the execution isn't immediately
finished, no.

* remove stupid test

I have no idea what these tests are about. They create system
workflows, but those aren't really system workflows, because these
aren't valid workflow_ids? And then the test asserts that the execution
exists? What's so special about that execution?

If it's trying to test that you can create system workflow with any
workflow_id, which I'm not sure it does, then no: you can't.
We validate the workflow_id to be on the whitelist, for system
worfklows.

* pass parameters in this test

* RD-1640 Group-aware queueing mechanism

When dequeueing executions, the order is
1. system workflows (if any, stop at that, and don't dequeue further)
2. same-deployment executions, but just one, because of course
   multiple can't run at the same time
3. other queued executions

When dequeueing in 3, only dequeue up to the group concurrency cap

* flake