RD-1640 and RD-1700 Bugfixes (#2788)
* RD-1700 Operation endpoints: make sure to use transactions

Use a transaction, and be really careful about getting the execution
object to make sure it's select with a WITH UPDATE.

Also, NOPTasks are not counted

* RD-1700 Don't count NOP operations

Some operations do start in an already TERMINATED state, so count them.
But, don't count NOP tasks. They aren't shown in the graph, no need
to show them here either.

* RD-1640 Dequeue correctly when a system execution finished

The `finished_execution.deployment.id` really needs to be under
the `if` :)
And of course, for no-deployment cases, `queued_executions` is
just that. Can't use a deployment id filter in there, duh.

* cancel_execution: take the execution object

Like everything else does now. Otherwise we just get an error when
passing the id into `generate_execution_token`