RD-1640 groups: execute up to the concurrency limit (#2791)
* When dequeueing executions, mark them as pending

Otherwise they can't run. Because exec-token doesn't match for queued
executions.

* Exec-group tests: cope with the new status checks

This one's boring

* RD-1640 exec-groups: execute up to the concurrency limit

* flake

* When creating executions, give them a status

Otherwise they'll never run.
Probably this status should just be defaulted somewhere, but for
right now, I'm going with just this

* exceute_system_workflow: exclude the current execution

When looking for running executions, obviously don't consider the
execution that we're trying to run.

* fix typo

* assert_no_snapshot: make the execution para optional

turns out we do sometimes want to run that check outside of the
context of an execution, eg. when uploading plugins