RD-1317 Labels for deployment-groups (#2824)
* Add a dep-groups labels model

* create depgroup labels in the PUT endpoint

when PUTting a depgroup, create labels for it as specified by the
request

* Pass the group's labels to deployments create-dep-env

Labels of the group are automatically applied to the deployments,
so pass them through.

* create-dep-env: deduplicate labels

For this, deduplicate them. So that if a label defined by the group
is the same as one defined by the deployments, it Just Works™,
as opposed to throwing an error

* add details to single-item depgroup endpoints

For the single-item (ie. NOT list) endpoints, add get-data, so that
labels are part of the response

* tests

* flake

* fix test docstrings

* also add asserts to that test

* normalize the type of labels

it needs to have get_labels_list called on it everywhere