inte-tests: use cloudify_internal_ca_cert.pem
we've made the external cert be signed (by default) by the "main"
CA (which is the internal CA), as opposed to self-signed.

This was because this was more friendly to the user, because now
they can just provide a single CA cert+key and have both certs signed
by that.

That change was in:
https://github.com/cloudify-cosmo/cloudify-manager-install/pull/1105/commits/b699a4edebb7bd8beac7b4064c707e4f758615af