exec-group start: only fetch the executions list once
This method used to access `self.execution` twice, in separate
transactions. This was then separate DB lookups, which could be
ordered differently. In other words,
`self.executions[:concurrency] + self.executions[concurrency:]` was
not necessarily the same thing (it would often be the same, because
postgres prefers sequential access, but not always).

This led to some executions being handled twice: they would get both
their status changed to queued, and be started. The start would then
throw the 401; and later they would be dequeued normally and succeed.

...and some executions would not be handled at all (because the
"twice" ones took their place) and would stay "pending" forever.

We could've added an `order by` to the collection, but a better
way is to just only fetch it once.