test_snapshot: use self.client for the restclient
making a new client is harder because utils.get_rest_client needs
more info than that: it needs to be ssl-enabled, so it needs the cert,
etc.

Instead, just avoid all that hassle by using the client we already
have, just switch the tenant as needed, using the util that already
exists.