test_cancel/resume group: wait for the deps to be created
these tests create a few deployments, so wait for them to be
created before doing anything with their executions.

also reduce the count to just 2, because if we are to wait for these
deployments, let's at not wait as long. Doesn't really matter for
this test anyway, can be 2 or 20 or 200.