node.plugins is a list, not a dict
default to the empty list, not the empty dict

Also rename the variable in dep-up. It's obvious that it's not a dict
there, because we immediately do .append and .remove on it.

Also change the gwetattr default in dep-up: I think it's actually
never used for .plugins, but might be for plugins_to_install

This fixes the inte-tests in test_deployment_update_misc.py