Create cascading executions in a transaction
Creating the executions must be in a transaction, because setting
deployment on the execution can load the blueprint from the
deployment, and that would flush the execution object into the db,
possibly before some of its required attributes were set.

Alternatively we could've used an `no_autoflush` block, but using
an explicit transaction allows us to create all the component
executions in a single transaction.

Note that every time we create & run an execution, creating the
execution object (or indeed thousands of them, in a group) can/should
be under a transaction, but the `.execute_workflow()` call itself
can't be, because we must make sure that the execution is saved into
the db before it hits rabbitmq (otherwise the mgmtworker could've
started executing it before it's been actually written).