_get_queued_executions: use db-side sorting for same-deployment
putting same-deployment executions first can be done by just sorting
them first in the query. No need for two queries.

Two queries were wrong because two concurrent executions of this
method (for different deployments) would want to lock the same rows,
but in a different order, and that leads to deadlocks.