RD-2046 Drop a query to avoid deadlocks (#2905)
* storage-manager: allow sort to have callable values

similar to what we did for filters. Sort conditions can be more
complex than just a simple ASC or DESC.

* _get_queued_executions: use db-side sorting for same-deployment

putting same-deployment executions first can be done by just sorting
them first in the query. No need for two queries.

Two queries were wrong because two concurrent executions of this
method (for different deployments) would want to lock the same rows,
but in a different order, and that leads to deadlocks.

* collapse the if

the two queries only differ by the sort

* _check_for_active_executions: also consider queued executions

if an execution for this deployment is already queued, why should
we be allowed to run? we also need to get queued.
(specifically, if a create-dep-env is currently queued, there's no
way we can run anything)

This creates a FIFO-per-deployment for executions

* _check_for_active_executions: compare on the deployment fk

avoids a join

* fix a typo