run dequeue_execution before starting it
rename the function to "refresh", and also run it in execute_workflow.
This is so that every start of an execution makes sure to use
the actual up-to-date deployment & parameters (wrt. race conditions
between starting executions/groups, and create-dep-envs finishing)

also rename the method because it obviously doesn't really "dequeue"