fix _check_for_active_executions logic
obviously we can't consider all queued executions to block the current
one, because then if we have 2 queued executions for a given
deployment, they will block each other.

Instead, only consider OLDER THAN CURRENT queued executions.

ie. if there's a create-dep-env queued, and you run install, the
install will be queued, because it'll notice the already-existing
create-dep-env. But when the create-dep-env gets dequeued, it will
be able to run, because the queued install is younger.