RD-2056 Reevaluate parameters when starting executions (#2909)
* execute-workflow: put parts of it under a transaction

that does multiple queries, no reason for this to not all be
under a single transaction

the workflow_executor.execute_workflow must not be in the same
transaction, because we must commit the execution before it
gets sent over to mgmtworker

* run dequeue_execution before starting it

rename the function to "refresh", and also run it in execute_workflow.
This is so that every start of an execution makes sure to use
the actual up-to-date deployment & parameters (wrt. race conditions
between starting executions/groups, and create-dep-envs finishing)

also rename the method because it obviously doesn't really "dequeue"

* get_queued_executions: move deployment logic into the db

checking for already-running deployments can be easily done on the
db side like this.

I've decided to not do it for the groups part (yet?) because that
made the query more complex than i'd like

* fix _check_for_active_executions logic

obviously we can't consider all queued executions to block the current
one, because then if we have 2 queued executions for a given
deployment, they will block each other.

Instead, only consider OLDER THAN CURRENT queued executions.

ie. if there's a create-dep-env queued, and you run install, the
install will be queued, because it'll notice the already-existing
create-dep-env. But when the create-dep-env gets dequeued, it will
be able to run, because the queued install is younger.

* flake