RD-1165 Storage backend for get_group_capability (#2943)
* RD-1165 implement get_group_capability

Storage backend for get_group_capability
Supports several variants of calling that function.

* Add test(s)

Only a single test, but with quite a few asserts to it. Checks all
the variants of get_group_capability.
All in a single test, because with the amount of setup it requires
(several deployments), it takes about 6 seconds anyway.