deployments + labels: several performance changes (#2950)
* delete_deployments_labels: use a sql DELETE

the orm is just too slow for this. Go down a level.

* similarly, remove_deployment_label_dependency

the get+delete is just a waste of time here, only db-delete instead

* remove_deployments_from_label_graph: fetch target deployment once

we passed just target_id to this function, and so we ended up fetching
that deployment multiple times. Instead, pass the deployment object
in, it's already fetched - once.

* start-queued-execs: use a single amqp connection

in start-queued, we run possibly multiple executions: only create
a single connection, and send them all using that, instead of
making a new connection per execution

* get-queued-executions: limit the query

avoid locking too many execution rows at once: have _some_ limit
on the query; 5 seeems fair for a first estimate

* add a missing mock

similar to mocking out the workflow-executor stuff in test-dep-groups,
mock it out here as well. No AMQP in these tests.