RD-2246 Dep-update: cleanup the unpacked new blueprint
This ports #2973 to 5.2.2

Dep-update unpacks and parses the new blueprint so that it can parse
the blueprint to figure out the differences.

However, it must also clean up, obviously!

Also, let's make it go under a subdirectory ('updated_blueprint')
in the deployment directory - that's way better than just removing
the deployment directory entirely: it could've been used for other
things as well.