RD-2250 Move several inte-tests from premium (#2978)
* move multi_tenancy tests here

c-premium's multi_tenancy inte-tests haven't been run in years,
so we're moving them here so that they can actually be run

* move _select to the base class

i give up, it's too useful, allow using it in other tests as well

it'll be used in one of the user tests

* bind create_rest_client to self

i don't want to keep repeating the host & ca_cert args. This is nicer.

* use self.rest_client in the multi_tenancy tests

* multi_tenancy tests: the status-reporter user doesnt exist anymore

* multi_tenancy tests: use empty_blueprint.yaml

those tests want "just a blueprint", and this one is cheaper to upload

* use self._select here

direct postgres access is a no-no

* also move test_authorization.py from cloudify-premium

* test_auth: rewrite test_change_role

it used to check if the user has admin rights by using .set_ssl,
but .set_ssl is not a thing anymore.
Instead, maintenance-mode works.

* test_auth: rewrite _test_user_in_tenant

- drop the test_ssl, because admin is checked in the other test anyway
- use a cheaper operation for testing non-viewer

* style

just free -lines

* flake

* move test_file_server_access

this is the last inte-tests file from c-premium!

* fix test_file_server_access.py

with the various changes we've been making

* FileServerAuth: strip leading resources/

nginx will pass resources/ as the beginning of the url, but the
list we're comparing the url to, doesn't have that. So strip it.