prepare_reset_storage: add a manager-side component
when we prepare the reset-storage script & config, also add some code
that runs manager-side, so that we can store and cache (in the
reset_storage_config.json file) some values that are known only to
the running manager: the user password hash, and the admin token.

This will allow the actual reset_storage.py script to just use
those stored values, instead of having to recompute them, which is
expensive.