inte-tests: performance and removing unused stuff (#2980)
* conftest: prettify rest_client/ca_cert

- use tmpdir_factory for ca_cert, not tempfile
- reuse that cert for the restclient

* s/assertEquals/assertEqual/

removes some warning

* s/assertRaisesRegexp/assertRaisesRegex/

also loosen the regex a bit

* storage_utils: allow providing an already-hashed password

not a big deal for the function, but I'm going to use it in tests,
to avoid the expensive pbkdf2 on every testcase

* prepare_reset_storage: add a manager-side component

when we prepare the reset-storage script & config, also add some code
that runs manager-side, so that we can store and cache (in the
reset_storage_config.json file) some values that are known only to
the running manager: the user password hash, and the admin token.

This will allow the actual reset_storage.py script to just use
those stored values, instead of having to recompute them, which is
expensive.

* reset-storage: use password_hash, not password

use the stored hash, instead of the password. This avoids computing
the hash again.

* reset-storage: use the pre-stored admin token

use the token we stored in the prepare script, instead of running
a script to compute it again. That script is extremely expensive to
run (1.5+ seconds) because it imports all of manager_rest

* move cleaning dirs to the reset_storage script

it's just faster to clean the dirs python-side than it is to run
rm -rf, and obviously also faster to run the rm -rf via docker exec

* remove several places that handled provider context

we barely even care about the provider context anymore. Only one test
still references it at all (test_blueprint_upload_bad_import_resolver),
and if that test doesnt restore it correctly, then this code didnt
do it either, because it didnt have anything about the import_resolver.

Instead, let's just remove this, and keep trying to forget that
provider_context was ever a thing

* remove disused function

* simplify the reset script config a bit

those things can just be inlined, no real reason to pass them

* test_executions: remove pointless calls

it is valid. Otherwise executions would never run.

* remove more unused things

* simplify the scheduled-for strftime

* remove unused files

the suites maybe was a nice idea, actually! too bad we didnt use it,
and it ended up getting outdated

* remove unused integration_tests_blueprint

man now was THAT old! Thankfully, it was super unused. The current
jenkins jobs have their own blueprints inside of cloudify-build-system

* re-add __init__.py here

unfortunately it's needed for now. I'll fix this one some other time.

* remove unused docl bullshit

* remove unused postgresql module

direct db access from outside the container is haram

* remove yet another unused function

* remove the license from a few init.py

it was the only thing there!!! what is the point, man!

* remove get_cfy

it was only used in that assert method, which was also unused!
it couldn't have worked anyway, because we (now) don't require an
outside-of-container cfy

* replace sh excepts with subprocess

these functions dont use sh, and so they dont throw sh, but they
instead use subprocess, and throw subprocess

* remove the last place that was using sh

it was not used anyway, lol

this allows us to drop sh

* remove yet another one, and flake

drop all the unused imports

and drop a function that couldnt even work, because that can't be
called from outside the container

* add an if-name-main to this script

this script's name starts with `test_` and so pytest picks it up
in an --collect-only. So let's make this script not have any
side-effects when imported.