remove test_zipped_plugin
I note that this test has no assertions, because .check just returns
a bool, doesnt assert.
And it doesn't pass, and I dont know what it was supposed to do
(hello_world??? should there be a plugin? it puts a targz
into uploaded-blueprints... should I be unpacking it?) so I'm just
going to drop this. I assure you blueprint upload works, because
otherwise haha.