rest tests: when running under xdist, create more databases!
this way, we can actually use xdist.
This makes an additional assumption about the db we're using for the
tests: now the user not only needs to be the owner of the primary
database we're using, it also needs to be a superuser. (well, only
when using xdist) (who cares, still the easiest and fastest-to-cleanup
way is to just the suggested `docker run`, I'm not going to let
these tests anywhere near a db that I'm not going to tear down after)