Simplify & speed up REST tests (#2990)
* remove test_zipped_plugin

I note that this test has no assertions, because .check just returns
a bool, doesnt assert.
And it doesn't pass, and I dont know what it was supposed to do
(hello_world??? should there be a plugin? it puts a targz
into uploaded-blueprints... should I be unpacking it?) so I'm just
going to drop this. I assure you blueprint upload works, because
otherwise haha.

* test_delete_blueprint: replace asserts with os.path.exists

instead of checking a fileserver (which normally is nginx, which
obviously doesn't exist in these tests), just os.path.exists the
uploaded file directly. It's let's say less nice, but it's worth it,
I promise.

* drop the idea of the fileserver

told you it was worth it. That was the ONLY test using the damn
fileserver.

* rest tests: when running under xdist, create more databases!

this way, we can actually use xdist.
This makes an additional assumption about the db we're using for the
tests: now the user not only needs to be the owner of the primary
database we're using, it also needs to be a superuser. (well, only
when using xdist) (who cares, still the easiest and fastest-to-cleanup
way is to just the suggested `docker run`, I'm not going to let
these tests anywhere near a db that I'm not going to tear down after)

* flake

* run -n 2 on jenkins

* toy with the limits a bit

* cache the query for checking if a db exists

this runs once for every class, no need to do it that often

* create_wheel: use a tempdir

gettempdir is just /tmp. Then it's just a constant path, which allows
interactions between tests

* dont split the restservice tests

whats the point???

* also xdist for the premium tests

* amqp-postgres tests: mock out rabbitmq

yeah it's worth it